Phase 4 batch 3: extract thread stream + context bar from SubmissionStatusView
Splits the chat-thread render (ThreadStream), the phase/heartbeat bar
above the composer (ThreadContextBar), and the shared heartbeat clock
(BuildHeartbeat) out of SubmissionStatusView.tsx, which drops from
1,599 to 1,264 lines. Presence-thought logic (PRESENCE_THOUGHT_MS,
presenceThought()) moves into its own presenceThought.ts so
BuildHeartbeat.tsx exports only the component, avoiding a
react-refresh/only-export-components warning.
